Added AuthMethod enum
